pyproject.toml
[tool.isort]
profile = 'black'
extend_skip = ['env']

[tool.black]
line-length = 180
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | \env
)/
'''

[tool.poetry]
name = "pydanticiati"
version = "0.1.0"
description = ""
authors = ["Joshua Brooks <josh.vdbroek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
geojson-pydantic = "^0.3.1"
fastapi-manage = "^0.8.0"
httpx = "^0.20.0"
pytest-asyncio = "^0.16.0"
pytest-httpx = "^0.14.0"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.10b0"
isort = "^5.9.3"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
